Implement inefficiently
Right now we do a linear scan to find the previous component for an
entity for add and remove, and we collect all components for each
entity. It should be possible to collect components only if requested by
the visitor. This kills performance in the following scenario:

n entities with n components each. Query one component at a time.
Should be n^2, but is n^3 as implemented.